requirements.txt
Scrapy==2.5.0
Unidecode==1.3.1
SQLAlchemy-1.4.23
